third_party/blink/web_tests/platform/generic/media/media-fragments/TC0076-expected.txt


Title: TC0076
Fragment: 't=0:60:00'
Comment: UA knows that this is an invalid media fragment, so it will play the entire media resource.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
END OF TEST
